+++ packages/react-native-renderer/src/ReactNativeRenderer.js
@@ -13,13 +13,13 @@ import type {ReactNodeList} from 'shared/ReactTypes';
 import './ReactNativeInjection';
 
 import * as ReactNativeFiberRenderer from 'react-reconciler/inline.native';
+// TODO: direct imports like some-package/src/* are bad. Fix me.
+import {getStackAddendumByWorkInProgressFiber} from 'react-reconciler/src/ReactDebugCurrentFiber';
 import * as ReactPortal from 'shared/ReactPortal';
 import * as ReactGenericBatching from 'events/ReactGenericBatching';
 import ReactVersion from 'shared/ReactVersion';
 // Module provided by RN:
 import UIManager from 'UIManager';
-
-import {getStackAddendumByWorkInProgressFiber} from 'shared/ReactFiberComponentTreeHook';
 
 import NativeMethodsMixin from './NativeMethodsMixin';
 import ReactNativeComponent from './ReactNativeComponent';
