+++ packages/react-reconciler/src/ReactDebugCurrentFiber.js
@@ -8,12 +8,53 @@
  */
 
 import {ReactDebugCurrentFrame} from 'shared/ReactGlobalSharedState';
-import {getStackAddendumByWorkInProgressFiber} from 'shared/ReactFiberComponentTreeHook';
+import {
+  IndeterminateComponent,
+  FunctionalComponent,
+  ClassComponent,
+  HostComponent,
+} from 'shared/ReactTypeOfWork';
+import describeComponentFrame from 'shared/describeComponentFrame';
 import getComponentName from 'shared/getComponentName';
 
 import type {Fiber} from './ReactFiber';
 
 type LifeCyclePhase = 'render' | 'getChildContext';
+
+function describeFiber(fiber: Fiber): string {
+  switch (fiber.tag) {
+    case IndeterminateComponent:
+    case FunctionalComponent:
+    case ClassComponent:
+    case HostComponent:
+      const owner = fiber._debugOwner;
+      const source = fiber._debugSource;
+      const name = getComponentName(fiber);
+      let ownerName = null;
+      if (owner) {
+        ownerName = getComponentName(owner);
+      }
+      return describeComponentFrame(name, source, ownerName);
+    default:
+      return '';
+  }
+}
+
+// This function can only be called with a work-in-progress fiber and
+// only during begin or complete phase. Do not call it under any other
+// circumstances. Note that this method works both in DEV and PROD.
+export function getStackAddendumByWorkInProgressFiber(
+  workInProgress: Fiber,
+): string {
+  let info = '';
+  let node = workInProgress;
+  do {
+    info += describeFiber(node);
+    // Otherwise this return pointer might point to the wrong tree:
+    node = node.return;
+  } while (node);
+  return info;
+}
 
 export let current: Fiber | null = null;
 export let phase: LifeCyclePhase | null = null;
